backend/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Datenbank
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
ldap3==2.9.1

# Redis
redis==5.2.1

# Validierung & Utils
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
email-validator==2.2.0

# Hintergrund-Tasks
celery==5.4.0

# Export
openpyxl==3.1.5

# Tests
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
